Move signed_edges_to_signed_nodes to util file
