+++ indra/explanation/model_checker/__init__.py
@@ -3,4 +3,3 @@ from .pysb import PysbModelChecker
 from .signed_graph import SignedGraphModelChecker
 from .unsigned_graph import UnsignedGraphModelChecker
 from .pybel import PybelModelChecker
-from .model_checker import signed_edges_to_signed_nodes, prune_signed_nodes
