+++ indra/explanation/model_checker/signed_graph.py
@@ -1,8 +1,7 @@
 import logging
 from . import ModelChecker
 from indra.statements import *
-from .model_checker import signed_edges_to_signed_nodes
-
+from ..pathfinding_util import signed_edges_to_signed_nodes
 
 logger = logging.getLogger(__name__)
